@@ -40,8 +40,10 @@
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
     "@types/react-test-renderer": "^16.8.2",
-    "@typescript-eslint/eslint-plugin": "^2.18.0",
-    "@typescript-eslint/parser": "^2.18.0",
+    "@types/scheduler": "^0.16.1",
+    "@types/use-subscription": "^1.0.0",
+    "@typescript-eslint/eslint-plugin": "^3.1.0",
+    "@typescript-eslint/parser": "^3.1.0",
     "@vuepress/plugin-pwa": "^1.2.0",
     "commitizen": "^4.0.3",
     "coveralls": "^3.0.5",
@@ -57,18 +59,22 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-react": "^7.14.2",
     "eslint-plugin-standard": "^4.0.0",
+    "experimental_react": "npm:react@experimental",
+    "experimental_react-dom": "npm:react-dom@experimental",
+    "experimental_react-test-renderer": "npm:react-test-renderer@experimental",
     "husky": "^3.0.1",
-    "jest": "^25.1.0",
+    "jest": "^26.0.1",
     "lerna": "^3.20.2",
     "prettier": "^1.18.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-test-renderer": "^16.12.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-test-renderer": "^16.13.1",
     "rxjs": "^6.5.4",
     "shx": "^0.3.2",
     "standard-version": "^6.0.1",
-    "ts-jest": "^25.0.0",
-    "typescript": "^3.5.3",
+    "ts-jest": "^26.1.0",
+    "typescript": "^3.9.5",
+    "use-subscription": "^1.4.1",
     "vuepress": "^1.2.0",
     "vuepress-plugin-container": "^2.1.2"
   }
