@@ -4,16 +4,14 @@ import { mapTo, take } from 'rxjs/operators'
 import { timer, merge } from 'rxjs'
 
 export const Warning: React.FC = () => {
-  const [isClose, close] = useObservableState<boolean, any>(inputs$ =>
-    merge(inputs$, timer(10000)).pipe(
-      take(1),
-      mapTo(true)
-    )
+  const [isClose, onClose] = useObservableState<boolean, React.MouseEvent>(
+    inputs$ => merge(inputs$, timer(10000)).pipe(take(1), mapTo(true)),
+    false
   )
 
   return isClose ? null : (
     <div className="notification is-info">
-      <button className="delete" onClick={close} />
+      <button className="delete" onClick={onClose} />
       If Components do not clean up properly in codesandbox, hit the refresh
       button above.
     </div>
