@@ -3,24 +3,24 @@ import { renderHook, act } from '@testing-library/react-hooks'
 import { of, Subject, BehaviorSubject } from 'rxjs'
 
 describe('useObservableGetState', () => {
-  it('should have the synced init value from Observable without rerendering', () => {
-    const outer$ = of(1, 2, 3)
-    const { result } = renderHook(() => useObservableGetState(outer$))
-    expect(result.current).toBe(3)
-  })
-
-  it('should not lose the init value from sync Observable when rerendering trigger from other props or states', () => {
-    const outer$ = of(1, 2, 3)
-    const { result, rerender } = renderHook(() => useObservableGetState(outer$))
-    expect(result.current).toBe(3)
-    rerender()
-    expect(result.current).toBe(3)
+  it('should start receiving values after first rendering', () => {
+    const spy = jest.fn()
+    const outer$ = of({ a: { b: 1 } }, { a: { b: 2 } }, { a: { b: 3 } })
+    const { result } = renderHook(() => {
+      const state = useObservableGetState(outer$, -1, 'a', 'b')
+      spy(state)
+      return state
+    })
+    expect(result.current).toEqual(3)
+    expect(spy).toHaveBeenCalledTimes(2)
+    expect(spy).toHaveBeenNthCalledWith(1, -1)
+    expect(spy).toHaveBeenNthCalledWith(2, 3)
   })
 
   it('should update value when the Observable emits value', () => {
     const outer$$ = new Subject<number>()
-    const { result } = renderHook(() => useObservableGetState(outer$$))
-    expect(result.current).toBeUndefined()
+    const { result } = renderHook(() => useObservableGetState(outer$$, 0))
+    expect(result.current).toBe(0)
 
     act(() => outer$$.next(1))
     expect(result.current).toBe(1)
@@ -31,7 +31,7 @@ describe('useObservableGetState', () => {
 
   it('should get value with 1-level path', () => {
     const outer$$ = new BehaviorSubject({ a: 'a' })
-    const { result } = renderHook(() => useObservableGetState(outer$$, 'a'))
+    const { result } = renderHook(() => useObservableGetState(outer$$, '', 'a'))
     expect(result.current).toBe('a')
 
     act(() => outer$$.next({ a: 'b' }))
@@ -41,7 +41,7 @@ describe('useObservableGetState', () => {
   it('should get value with 3-level path', () => {
     const outer$$ = new BehaviorSubject({ a: { b: { c: 'a' } } })
     const { result } = renderHook(() =>
-      useObservableGetState(outer$$, 'a', 'b', 'c')
+      useObservableGetState(outer$$, '', 'a', 'b', 'c')
     )
     expect(result.current).toBe('a')
 
