@@ -2,10 +2,10 @@ export { useObservable } from './use-observable'
 export { useLayoutObservable } from './use-layout-observable'
 export { useObservableCallback } from './use-observable-callback'
 export { useSubscription } from './use-subscription'
+export { useLayoutSubscription } from './use-layout-subscription'
 export { useObservableState } from './use-observable-state'
 export { useObservableGetState } from './use-observable-get-state'
 export { useObservablePickState } from './use-observable-pick-state'
-export { useObservablePropsCallback } from './use-observable-props-callback'
 export { useObservableSuspense } from './use-observable-suspense'
 export { ObservableResource } from './observable-resource'
 export {
