@@ -1,35 +1,20 @@
 import { Observable, Subscription } from 'rxjs'
-import { useForceUpdate } from './helpers'
-import { useEffect, useRef } from 'react'
+import { useForceUpdate, useIsomorphicLayoutEffect } from './helpers'
+import { useRef } from 'react'
+
+// I know this is copy-paste from './useSubscription.ts'.
+// Instead of leaving the performance penalty to users,
+// I'll keep it this way until it gets too hard to maintain.
 
 /**
- * Accepts an Observable and optional `next`, `error`, `complete` functions.
- * These functions must be in correct order.
- * Use `undefined` or `null` for placeholder.
+ * Same as [[useSubscription]] except the subscription is established
+ * under `useLayoutEffect`.
  *
- * Subscription will unsubscribe when unmount, you can also
- * unsubscribe manually.
+ * Useful when values are needed before DOM paint.
  *
- * ⚠ **Note:** To make it concurrent mode compatible, the subscription happens
- * after the render is committed to the screen
- * which means even the Observable emits synchronous values
- * they will arrive after the first rendering.
- *
- * Note that changes of callbacks will not trigger
- * an emission. If you need that just create another
- * Observable of the callback with [[useObservable]].
- *
- * (From v2.0) You can access closure directly inside callback like in `useEffect`.
- * `useSubscription` will ensure the latest callback is called.
- *
- * (From v2.3.4) when the Observable changes `useSubscription` will automatically
- * unsubscribe the old one and resubscribe to the new one.
- *
- * ⚠ **Note:** Due to the design of RxJS, once an error occurs in an observable, the observable
- * is killed.
- * You should prevent errors from reaching observables or `catchError` in sub-observables.
- * You can also make the observable as state and replace it on error.
- * `useSubscription` will automatically switch to the new one.
+ * Use it scarcely as it runs synchronously before browser paint.
+ * Too many synchronous emissons from the observable could
+ * stretch the commit phase.
  *
  * @template TInput Input value within Observable.
  *
@@ -38,13 +23,13 @@ import { useEffect, useRef } from 'react'
  * @param error Notify when a new error is thrown.
  * @param complete Notify when the Observable is complete.
  */
-export function useSubscription<TInput>(
+export function useLayoutSubscription<TInput>(
   input$: Observable<TInput>,
   next?: ((value: TInput) => void) | null | undefined,
   error?: ((error: any) => void) | null | undefined,
   complete?: (() => void) | null | undefined
 ): React.MutableRefObject<Subscription | undefined>
-export function useSubscription<TInput>(
+export function useLayoutSubscription<TInput>(
   ...args: [
     Observable<TInput>,
     ((value: TInput) => void) | null | undefined,
@@ -60,7 +45,7 @@ export function useSubscription<TInput>(
   const subscriptionRef = useRef<Subscription>()
   const errorRef = useRef<Error | null>()
 
-  useEffect(() => {
+  useIsomorphicLayoutEffect(() => {
     errorRef.current = null
 
     // keep in closure for checking staleness
@@ -68,6 +53,10 @@ export function useSubscription<TInput>(
 
     const subscription = input$.subscribe({
       next: value => {
+        // layout effect runs synchronously
+        // this should never hit
+        // keeping for alignment to useEffect
+        /* istanbul ignore if */
         if (input$ !== argsRef.current[0]) {
           // stale observable
           return
@@ -77,6 +66,10 @@ export function useSubscription<TInput>(
         }
       },
       error: error => {
+        // layout effect runs synchronously
+        // this should never hit
+        // keeping for alignment to useEffect
+        /* istanbul ignore if */
         if (input$ !== argsRef.current[0]) {
           // stale observable
           return
@@ -89,6 +82,10 @@ export function useSubscription<TInput>(
         forceUpdate()
       },
       complete: () => {
+        // layout effect runs synchronously
+        // this should never hit
+        // keeping for alignment to useEffect
+        /* istanbul ignore if */
         if (input$ !== argsRef.current[0]) {
           // stale observable
           return
