@@ -1,24 +1,32 @@
+import { useDebugValue } from 'react'
 import { Observable } from 'rxjs'
 import { map, distinctUntilChanged } from 'rxjs/operators'
 import { useObservableState } from './use-observable-state'
 import { useObservable } from './use-observable'
 
 /**
- * Creates an object composed of the picked state properties.
+ * Creates an object composed of the picked state properties. Similar to lodash `pick`.
  * Changes of any of these properties will trigger a rerendering.
  * Errors are thrown on unreachable path.
  *
  * @param state$ Output state.
  * @param keys keys of state
  */
-export function useObservablePickState<
-  TState,
-  TKeys extends Array<keyof TState>
->(
+export function useObservablePickState<TState, TKeys extends keyof TState>(
   state$: Observable<TState>,
-  ...keys: TKeys
-): { [K in TKeys[number]]: TState[K] } | undefined {
-  return useObservableState(
+  initialState:
+    | (TKeys extends never ? TState : { [K in TKeys]: TState[K] })
+    | (() => TKeys extends never ? TState : { [K in TKeys]: TState[K] }),
+  ...keys: TKeys[]
+): TKeys extends never ? TState : { [K in TKeys]: TState[K] }
+export function useObservablePickState<TState, TKeys extends keyof TState>(
+  state$: Observable<TState>,
+  initialState:
+    | { [K in TKeys]: TState[K] }
+    | (() => { [K in TKeys]: TState[K] }),
+  ...keys: TKeys[]
+): { [K in TKeys]: TState[K] } {
+  const value = useObservableState(
     useObservable(() =>
       state$.pipe(
         distinctUntilChanged((s1, s2) => keys.every(k => s1[k] === s2[k])),
@@ -26,10 +34,13 @@ export function useObservablePickState<
           keys.reduce(
             // eslint-disable-next-line no-sequences
             (o, k) => ((o[k] = state[k]), o),
-            {} as { [K in TKeys[number]]: TState[K] }
+            {} as { [K in TKeys]: TState[K] }
           )
         )
       )
-    )
+    ),
+    initialState
   )
+  useDebugValue(value)
+  return value
 }
