@@ -1,3 +1,4 @@
+import { useDebugValue, useState } from 'react'
 import { useSubscription } from './use-subscription'
 import { useForceUpdate } from './helpers'
 import { ObservableResource } from './observable-resource'
@@ -13,7 +14,20 @@ import { ObservableResource } from './observable-resource'
 export function useObservableSuspense<TInput, TOutput extends TInput = TInput>(
   resource: ObservableResource<TInput, TOutput>
 ): TOutput {
+  const resourceValue = resource.read()
+  const [state, setState] = useState<TOutput>(resourceValue)
   const forceUpdate = useForceUpdate()
-  useSubscription(resource.shouldUpdate$$, forceUpdate)
-  return resource.read()
+
+  useSubscription(resource.shouldUpdate$$, valueRef => {
+    // ObservableResource supports Stale-While-Revalidate pattern.
+    // Schedule states to prevent tearing.
+    if (valueRef) {
+      setState(valueRef.current)
+    } else {
+      forceUpdate()
+    }
+  })
+
+  useDebugValue(state)
+  return state
 }
